[REFACTOR] DetailView 코드 분리 (#64)
