+++ Manito/Manito/Screens/Detail-Wait/DetailWaitViewController.swift
@@ -234,7 +234,7 @@ class DetailWaitViewController: BaseViewController {
     }
 
     private func presentModal() {
-        let modalViewController = DetailModalController()
+        let modalViewController = DetailEditViewController()
 
         present(modalViewController, animated: true, completion: nil)
     }
